PR cahnges + header fixes
