+++ components/CoinProfiling/styles.ts
@@ -9,7 +9,7 @@ export const main = css({
   });
 
 export const headerWrapper = css({
-    position: "fixed",
+    // position: "fixed",
     top: 0,
     left: 0,
     width: "100%",
