+++ components/HomePage/Index.tsx
@@ -1,0 +1,11 @@
+import * as styles from "./styles"
+
+const HomePage = () => {
+    return (
+        <div css={styles.container}>
+
+        </div>
+    );
+}
+
+export default HomePage;
