+++ components/HomePage/styles.ts
@@ -1,0 +1,6 @@
+import { css } from "@emotion/react";
+import { colors } from "zebpay-ui";
+
+export const container = css({
+    backgroundColor: colors.Zeb_BG_Blue
+});
