+++ constants/api.ts
@@ -1,2 +1,3 @@
-export const COINMARKETCAP_LATEST = "https://pro-api.coinmarketcap.com/v2/cryptocurrency/info";
-export const COINMARKETCAP_INFO = "https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest";
+export const COINMARKETCAP_INFO = "https://pro-api.coinmarketcap.com/v2/cryptocurrency/info";
+export const COINMARKETCAP_LATEST = "https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest";
+
