+++ next.config.mjs
@@ -2,6 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   basePath: "/coin-profiling",
+  images: {
+    domains: ["s2.coinmarketcap.com"], 
+  }
 };
 
 export default nextConfig;
