+++ public/images/index.ts
@@ -54,9 +54,7 @@ import ic_category from "./icons/category.svg";
 import ic_date from "./icons/date.svg";
 import ic_duration from "./icons/duration.svg";
 import ic_google_play from "./icons/google_play.svg";
-//import ic_mail from "./icons/mail.svg";
 import ic_no_blogs_found from "./icons/no_blogs_found.svg";
-//import ic_seperator from "./icons/seperator.svg";
 import ic_subscribed from "./icons/subscribed.svg";
 import ic_views from "./icons/views.svg";
 import ic_zebpay from "./icons/zebpay.svg"; 
@@ -83,7 +81,6 @@ import ic_cross from "./Icons/cross.svg";
 import ic_investmentIcon from "./illustrations/investmentIcon.png";
 import ic_no_search from "./illustrations/no-search.png";
 import ic_tb_gradient from './illustrations/trading_banner_gradient.png'
-//import ic_no_search from "./illustrations/no-search.png";
 
 const AssetsImg = {
   i_zebpay_logo,
@@ -157,9 +154,7 @@ const AssetsImg = {
   ic_date,
   ic_duration,
   ic_google_play,
-  //ic_mail,
   ic_no_blogs_found,
-  //ic_seperator,
   ic_subscribed,
   ic_views,
   ic_zebpay,
@@ -168,7 +163,6 @@ const AssetsImg = {
   ic_cross_blue,
   ic_clock_white,
   ic_category_white,
-  //ic_no_search
 };
 
 export default AssetsImg;
